+++ Markdown/Tokens/TokenSegment.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Markdown.Extensions;
 
 namespace Markdown
 {
